Day4

+++ src/events/events.controller.ts
@@ -1,7 +1,9 @@
 import {
   Body,
+  ClassSerializerInterceptor,
   Controller,
   Delete,
+  ForbiddenException,
   Get,
   HttpCode,
   Logger,
@@ -11,32 +13,30 @@ import {
   Patch,
   Post,
   Query,
+  SerializeOptions,
+  UseGuards,
+  UseInterceptors,
   UsePipes,
   ValidationPipe,
 } from '@nestjs/common';
-import { InjectRepository } from '@nestjs/typeorm';
-import { Like, MoreThan, Repository } from 'typeorm';
-import { Attendee } from './attendee.entity';
-import { Event } from './event.entity';
+import { AuthGuardJwt } from 'src/auth/auth-guard.jwt';
+import { CurrentUser } from 'src/auth/current-user.decorator';
+import { User } from 'src/auth/user.entity';
+import { EventsService } from './event.service';
 import { CreateEventDto } from './input/create-event.dto';
 import { ListEvents } from './input/list.events';
 import { UpdateEventDto } from './input/update-event.dto';
-import { EventsService } from './event.service';
 
 @Controller('/events')
+@SerializeOptions({ strategy: 'excludeAll' })
 export class EventsController {
   private readonly logger = new Logger(EventsController.name);
 
-  constructor(
-    @InjectRepository(Event)
-    private readonly repository: Repository<Event>,
-    @InjectRepository(Attendee)
-    private readonly attendeeRepository: Repository<Attendee>,
-    private readonly eventsService: EventsService,
-  ) {}
+  constructor(private readonly eventsService: EventsService) {}
 
   @Get()
   @UsePipes(new ValidationPipe({ transform: true }))
+  @UseInterceptors(ClassSerializerInterceptor)
   async findAll(@Query() filter: ListEvents) {
     const events =
       await this.eventsService.getEventsWithAttendeeCountFilteredPaginated(
@@ -50,60 +50,57 @@ export class EventsController {
     return events;
   }
 
-  @Get('/practice')
-  async practice() {
-    return await this.repository.find({
-      select: ['id', 'when'],
-      where: [
-        {
-          id: MoreThan(3),
-          when: MoreThan(new Date('2021-02-12T13:00:00')),
-        },
-        {
-          description: Like('%meet%'),
-        },
-      ],
-      take: 2,
-      order: {
-        id: 'DESC',
-      },
-    });
-  }
+  // @Get('/practice')
+  // async practice() {
+  //   // return await this.repository.find({
+  //   //   select: ['id', 'when'],
+  //   //   where: [{
+  //   //     id: MoreThan(3),
+  //   //     when: MoreThan(new Date('2021-02-12T13:00:00'))
+  //   //   }, {
+  //   //     description: Like('%meet%')
+  //   //   }],
+  //   //   take: 2,
+  //   //   order: {
+  //   //     id: 'DESC'
+  //   //   }
+  //   // });
+  // }
 
-  @Get('practice2')
-  async practice2() {
-    // // return await this.repository.findOne(
-    // //   1,
-    // //   { relations: ['attendees'] }
-    // // );
-    // const event = await this.repository.findOne(
-    //   1,
-    //   { relations: ['attendees'] }
-    // );
-    // // const event = new Event();
-    // // event.id = 1;
+  // @Get('practice2')
+  // async practice2() {
+  //   // // return await this.repository.findOne(
+  //   // //   1,
+  //   // //   { relations: ['attendees'] }
+  //   // // );
+  //   // const event = await this.repository.findOne(
+  //   //   1,
+  //   //   { relations: ['attendees'] }
+  //   // );
+  //   // // const event = new Event();
+  //   // // event.id = 1;
 
-    // const attendee = new Attendee();
-    // attendee.name = 'Using cascade';
-    // // attendee.event = event;
+  //   // const attendee = new Attendee();
+  //   // attendee.name = 'Using cascade';
+  //   // // attendee.event = event;
 
-    // event.attendees.push(attendee);
-    // // event.attendees = [];
+  //   // event.attendees.push(attendee);
+  //   // // event.attendees = [];
 
-    // // await this.attendeeRepository.save(attendee);
-    // await this.repository.save(event);
+  //   // // await this.attendeeRepository.save(attendee);
+  //   // await this.repository.save(event);
 
-    // return event;
+  //   // return event;
 
-    return await this.repository
-      .createQueryBuilder('e')
-      .select(['e.id', 'e.name'])
-      .orderBy('e.id', 'ASC')
-      .take(3)
-      .getMany();
-  }
+  //   // return await this.repository.createQueryBuilder('e')
+  //   //   .select(['e.id', 'e.name'])
+  //   //   .orderBy('e.id', 'ASC')
+  //   //   .take(3)
+  //   //   .getMany();
+  // }
 
   @Get(':id')
+  @UseInterceptors(ClassSerializerInterceptor)
   async findOne(@Param('id', ParseIntPipe) id: number) {
     // console.log(typeof id);
     const event = await this.eventsService.getEvent(id);
@@ -119,37 +116,55 @@ export class EventsController {
   // It can be done per method, or for every method when you
   // add it at the controller level.
   @Post()
-  async create(@Body() input: CreateEventDto) {
-    return await this.repository.save({
-      ...input,
-      when: new Date(input.when),
-    });
+  @UseGuards(AuthGuardJwt)
+  @UseInterceptors(ClassSerializerInterceptor)
+  async create(@Body() input: CreateEventDto, @CurrentUser() user: User) {
+    return await this.eventsService.createEvent(input, user);
   }
 
   // Create new ValidationPipe to specify validation group inside @Body
   // new ValidationPipe({ groups: ['update'] })
   @Patch(':id')
-  async update(@Param('id') id, @Body() input: UpdateEventDto) {
-    const event = await this.repository.findOne(id);
+  @UseGuards(AuthGuardJwt)
+  @UseInterceptors(ClassSerializerInterceptor)
+  async update(
+    @Param('id') id,
+    @Body() input: UpdateEventDto,
+    @CurrentUser() user: User,
+  ) {
+    const event = await this.eventsService.getEvent(id);
 
     if (!event) {
       throw new NotFoundException();
     }
 
-    return await this.repository.save({
-      ...event,
-      ...input,
-      when: input.when ? new Date(input.when) : event.when,
-    });
+    if (event.organizerId !== user.id) {
+      throw new ForbiddenException(
+        null,
+        `You are not authorized to change this event`,
+      );
+    }
+
+    return await this.eventsService.updatedEvent(event, input);
   }
 
   @Delete(':id')
+  @UseGuards(AuthGuardJwt)
   @HttpCode(204)
-  async remove(@Param('id') id) {
-    const result = await this.eventsService.deleteEvent(id);
+  async remove(@Param('id') id, @CurrentUser() user: User) {
+    const event = await this.eventsService.getEvent(id);
 
-    if (result?.affected !== 1) {
+    if (!event) {
       throw new NotFoundException();
     }
+
+    if (event.organizerId !== user.id) {
+      throw new ForbiddenException(
+        null,
+        `You are not authorized to remove this event`,
+      );
+    }
+
+    await this.eventsService.deleteEvent(id);
   }
 }
